backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==1.10.15
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.0.8
python-multipart==0.0.9
boto3==1.35.15
httpx==0.27.2
tenacity==8.5.0
faiss-cpu==1.7.4
fpdf2==2.7.9
python-dotenv==1.0.1
numpy==1.26.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
